[issue#4] Fix Oomph Setup for new repository structure
With the new setup it is possible to select only a subset of the
projects. Each repository is a subproject. It is also possible to select
one project first and install the dev env, and later execute Oomph and
install another. Therefore select the "Overwrite" flag on the final
wizard page.

Buildship 2.0 is added to the installed tools and replaces STS Gradle
feature. Gradle metadata needs to be checked in for the Project Import
action. Otherwise import fails with exception and the setup process
stops.

A Working Set is defined per repository, except for Eclipse. Since the
xtext-eclipse project has quite many projects, I devided them into
"Eclipse", "Eclipse Features" and "Eclipse Examples". A Working Set
"Root Projects" collects the Build root projects of all repositories.

Java version is set to 1.8 in JRE Task and org.eclipse.jdt.core
preferences.

Tycho m2e connectors installed with p2 task.

Target Platform and API Baseline are installed with siubproject "Eclipse
Support".